+++ examples/datasets/spm_faces_dataset.py
@@ -39,7 +39,7 @@ raw_fname = data_path + '/MEG/spm/SPM_CTF_MEG_example_faces%d_3D.ds'
 raw = io.read_raw_ctf(raw_fname % 1, preload=True)  # Take first run
 # Here to save memory and time we'll downsample heavily -- this is not
 # advised for real data as it can effectively jitter events!
-raw.resample(120., npad='auto')
+raw.resample(120., method='poly')
 
 picks = mne.pick_types(raw.info, meg=True, exclude='bads')
 raw.filter(1, 30, method='fir', fir_design='firwin')